+++ docs/history.md
@@ -4,6 +4,7 @@
 * Adding enable / disable sprint config
 * Adding enable / disable dig (!caution the bot can stucks)
 * Adding reloading configuration button
+* Unify botWebsocket.on(XX events) to botWebsocket.on('action')
 
 
 ## 0.1.6